+++ src/enums/dbServiceErrorEnum.js
@@ -1,1 +1,2 @@
 exports.QUERY_FAILED = 'failed to query xp database'
+exports.ACCOUNT_NOT_FOUND = 'account not found'