+++ src/index.js
@@ -4,6 +4,8 @@ const queryHandler = require('./handlers/queryHandler.js')
 const queryErrorHandler = require('./handlers/queryErrorHandler.js')
 const depositHandler = require('./handlers/depositHandler.js')
 const depositErrorHandler = require('./handlers/depositErrorHandler.js')
+const balanceHandler = require('../src/handlers/balanceHandler.js')
+const balanceErrorHandler = require('../src/handlers/balanceErrorHandler.js')
 
 exports.handler = async (event, context) => {
     logEvent(event);
@@ -27,6 +29,14 @@ exports.handler = async (event, context) => {
             await depositHandler.handle(event)
         } catch (err) {
             const errorHandlerResp = await depositErrorHandler.handle(err)
+            statusCode = errorHandlerResp.statusCode
+            body = errorHandlerResp.body
+        }
+    } else if (event.requestContext.routeKey == 'GET /xp/balance') {
+        try {
+            body = await balanceHandler.handle(event)
+        } catch (err) {
+            const errorHandlerResp = await balanceErrorHandler.handle(err)
             statusCode = errorHandlerResp.statusCode
             body = errorHandlerResp.body
         }