+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This project attempts to conform to [SemVer](https://semver.org/) guidelines whe
 
 ## [v0.2.0]
 * [Implementing Deposit XP API](https://github.com/Ubunfu/mc-xp-bank/pull/1)
+* [Implementing XP Account Balance API](https://github.com/Ubunfu/mc-xp-bank/pull/2)
 
 ## [v0.1.0]
 * Implementing Query XP API